be/alumninet/requirements.txt
Django==4.2.1
gunicorn==20.1.0
djangorestframework==3.14.0
drf-jwt==1.19.2
whitenoise==6.4.0
shortuuid==1.0.11
Pillow==10.1.0
requests==2.31.0
thefuzz==0.22.1
django-prometheus==2.3.1
pika==1.3.2
django-redis==5.2.0